cover SyntaxErrorInQueryMethodRuleTest on pgsql
